fix: Storage Package
The bucket storage object had a context set to canel in 10 seconds, which
was too short.

You also could not set it in the New method for initializing a bucket as
storage. The only work-a-round was to initialize it then set the duration
on the structure returned.

This commit allows you to set it as paramter in its new function, though it
breaks the current API, this is going in as a fix instead of a breaking change;
As it was easy to miss and resulted unexpected behavior for developers.