rmv: GetContextWithTimeout
Removed from the www package the GetContextWithTimeout function, as it
is was too simple and did not warrant its own function. The function simply
created a context with a timeout, which can be done directly where needed.

BREAKING CHANGE: Code that used GetContextWithTimeout must now create contexts
with timeouts directly using context.WithTimeout.