fix: Build Warnings for Log Handler
Some calls to log methods were missing the first parameter which should
be a string representing a template for the log message. This caused
build warnings about missing arguments "non-constant format string in
call". This commit adds the missing string parameters to the log calls
to resolve the warnings and ensure that the log messages are properly
formatted.