fix: Backend Templates Variable Values
Can now be any type instead of just a string.

It previously only allowed template variables to have a string value.
The original intension was never meant to limit it to just this type,
it was an oversite is all. This corrects that withtou introducing a
break in the API. Though just strings are safter; this allows a
greater depth and removes undue limitation to allow complex Templates.
While you MUST stay vigilent with security, that should go without
saying.